-알림 게시물이 삭제 되었을 때 알림도 함께 삭제 -게시물 지워지면 알림 지우기 - 재학 증명서 제출하지 않은 사람은 신고 알림 못 받음

+++ src/main/java/com/spring/ion/yjw/service/FlagService.java
@@ -91,9 +91,7 @@ public class FlagService {
     }
 
 
-
-
-    public void delete(int id) {
+    public void delete(long id) {
         flagRepository.delete(id);
     }
